adq patch - refactor economy : extract weekly optimization classes in separate source files
